+++ angular_components/lib/material_select/material_dropdown_select_accessor.dart
@@ -23,14 +23,14 @@ import 'material_dropdown_select.dart';
 )
 class DropdownSelectValueAccessor<T> extends BaseDropdownSelectValueAccessor<T>
     implements ControlValueAccessor, OnDestroy {
-  StreamSubscription _selectionChangesSub;
+  StreamSubscription? _selectionChangesSub;
   DropdownSelectValueAccessor(MaterialDropdownSelectComponent select)
-      : super(select, SelectionModel.single());
+      : super(select as MaterialDropdownSelectComponent<T*>, SelectionModel.single());
 
   @override
   void registerOnChange(callback) {
     _selectionChangesSub = selectionModel.selectionChanges.listen((_) {
-      var value = (selectionModel.selectedValues == null ||
+      T? value = (selectionModel.selectedValues == null ||
               selectionModel.selectedValues.isEmpty)
           ? null
           : selectionModel.selectedValues?.first;
@@ -67,9 +67,9 @@ class DropdownSelectValueAccessor<T> extends BaseDropdownSelectValueAccessor<T>
   ],
 )
 class MultiDropdownSelectValueAccessor<T>
-    extends BaseDropdownSelectValueAccessor<T>
+    extends BaseDropdownSelectValueAccessor<T?>
     implements ControlValueAccessor, OnDestroy {
-  StreamSubscription<List<SelectionChangeRecord<T>>> selectionChangesSub;
+  StreamSubscription<List<SelectionChangeRecord<T?>>>? selectionChangesSub;
 
   MultiDropdownSelectValueAccessor(MaterialDropdownSelectComponent<T> select)
       : super(select, MultiSelectionModel<T>());
@@ -106,7 +106,7 @@ abstract class BaseDropdownSelectValueAccessor<T>
   final MaterialDropdownSelectComponent<T> _select;
   @protected
   final SelectionModel<T> selectionModel;
-  StreamSubscription _visibileSub;
+  StreamSubscription? _visibileSub;
   bool initialized = false;
 
   BaseDropdownSelectValueAccessor(this._select, this.selectionModel);
@@ -124,7 +124,7 @@ abstract class BaseDropdownSelectValueAccessor<T>
   void registerOnTouched(callback) {
     _visibileSub = _select.visibleStream.listen((_) {
       // We only need the first event. Cancel the subscription.
-      _visibileSub.cancel();
+      _visibileSub!.cancel();
       callback();
     });
   }