+++ angular_components/lib/model/selection/locked_limit_selection_options.dart
@@ -23,16 +23,16 @@ class LockedLimitSelectionOptions<T> extends DelegatingSelectionOptions<T>
   final Disposer _disposer = Disposer.oneShot();
 
   bool _unlockLimit = false;
-  int _currentLimit;
+  int? _currentLimit;
   bool _hasMoreOptions = true;
 
   /// The cached [OptionGroup]s from filtering with the [currentQuery].
   /// This is used to manipulate the underlying OptionGroups without
   /// accidentally throwing away any of the options.
-  List<OptionGroup<T>> _filteredGroups;
+  List<OptionGroup<T>>? _filteredGroups;
 
   /// The flattened [_filteredGroups].
-  List<T> _flattenedOptions;
+  List<T>? _flattenedOptions;
 
   LockedLimitSelectionOptions(SelectionOptions<T> options, this.lockedLimit)
       : _options = options,
@@ -55,26 +55,28 @@ class LockedLimitSelectionOptions<T> extends DelegatingSelectionOptions<T>
   bool get hasMoreOptions => _hasMoreOptions;
 
   @override
-  Object get currentQuery => (_options as Filterable).currentQuery;
+  Object? get currentQuery => (_options as Filterable).currentQuery;
 
   @override
-  int get currentLimit => _currentLimit;
+  int? get currentLimit => _currentLimit;
 
   @override
-  List<OptionGroup<T>> get optionGroups => _filteredGroups;
+  List<OptionGroup<T>> get optionGroups => _filteredGroups ?? [];
 
   @override
-  List<T> get optionsList => _flattenedOptions;
+  List<T> get optionsList => _flattenedOptions ?? [];
 
   /// Filter up to [limit] results using [filterQuery]. [limit] is used instead
   /// of [lockedLimit] iff [unlockLimit] is true.
   @override
-  DisposableFuture<bool> filter(Object filterQuery, {int limit = UNLIMITED}) {
+  DisposableFuture<bool> filter(Object filterQuery, {int? limit = UNLIMITED}) {
     _currentLimit = _unlockLimit ? limit : lockedLimit;
     // If the limit is locked and finite, only filter one extra item to
     // determine if results will be truncated to meet the lockedLimit.
-    if (!(_unlockLimit || limit == UNLIMITED)) {
-      limit++;
+    if (limit != null) {
+      if (!(_unlockLimit || limit == UNLIMITED)) {
+        limit++;
+      }
     }
 
     DisposableFuture<bool> filtered =
@@ -99,22 +101,22 @@ class LockedLimitSelectionOptions<T> extends DelegatingSelectionOptions<T>
     } else {
       _setLimitedOptions(_options.optionGroups, _currentLimit);
     }
-    _hasMoreOptions = _flattenedOptions.length < _options.optionsList.length;
+    _hasMoreOptions = _flattenedOptions!.length < _options.optionsList.length;
   }
 
-  void _setLimitedOptions(List<OptionGroup<T>> groups, int maxItems) {
+  void _setLimitedOptions(List<OptionGroup<T>> groups, int? maxItems) {
     int totalOptionCount = 0;
     _filteredGroups = [];
     for (OptionGroup<T> group in groups) {
-      if (totalOptionCount + group.length < maxItems) {
-        _filteredGroups.add(group);
+      if (totalOptionCount + group.length < maxItems!) {
+        _filteredGroups!.add(group);
         totalOptionCount += group.length;
       } else {
-        var lastOptions = group.take(maxItems - totalOptionCount).toList();
-        _filteredGroups.add(OptionGroup<T>(lastOptions));
+        List<T> lastOptions = group.take(maxItems - totalOptionCount).toList();
+        _filteredGroups!.add(OptionGroup<T>(lastOptions));
         break;
       }
     }
-    _flattenedOptions = _filteredGroups.expand((i) => i).toList();
+    _flattenedOptions = _filteredGroups!.expand((i) => i).toList();
   }
 }