+++ angular_components/lib/src/model/selection/selection_change_notifier.dart
@@ -23,18 +23,16 @@ abstract class SelectionObservable<T> {
 
 /// Mixin for providing [SelectionModel.selectionChanges].
 abstract class SelectionChangeNotifier<T> implements SelectionModel<T> {
-  StreamController<List<SelectionChangeRecord<T>>> _selectionChangeController;
-  List<SelectionChangeRecord<T>> _selectionChangeRecords;
+  StreamController<List<SelectionChangeRecord<T>>>? _selectionChangeController;
+  List<SelectionChangeRecord<T>> _selectionChangeRecords = [];
 
   @override
   bool deliverSelectionChanges() {
-    if (hasSelectionObservers &&
-        _selectionChangeRecords != null &&
-        _selectionChangeRecords.isNotEmpty) {
+    if (hasSelectionObservers && _selectionChangeRecords.isNotEmpty) {
       var records = UnmodifiableListView<SelectionChangeRecord<T>>(
           _selectionChangeRecords);
-      _selectionChangeRecords = null;
-      _selectionChangeController.add(records);
+      _selectionChangeRecords.clear();
+      _selectionChangeController!.add(records);
       return true;
     } else {
       return false;
@@ -46,8 +44,8 @@ abstract class SelectionChangeNotifier<T> implements SelectionModel<T> {
       {Iterable<T> added = const [], Iterable<T> removed = const []}) {
     if (hasSelectionObservers) {
       var record = SelectionChangeRecord<T>(added: added, removed: removed);
-      if (_selectionChangeRecords == null) {
-        _selectionChangeRecords = [];
+      if (_selectionChangeRecords.isEmpty) {
+        //_selectionChangeRecords = [];
         scheduleMicrotask(deliverSelectionChanges);
       }
       _selectionChangeRecords.add(record);
@@ -57,7 +55,7 @@ abstract class SelectionChangeNotifier<T> implements SelectionModel<T> {
   @override
   bool get hasSelectionObservers {
     return _selectionChangeController != null &&
-        _selectionChangeController.hasListener;
+        _selectionChangeController!.hasListener;
   }
 
   @override
@@ -67,7 +65,7 @@ abstract class SelectionChangeNotifier<T> implements SelectionModel<T> {
           StreamController<List<SelectionChangeRecord<T>>>.broadcast(
               sync: true);
     }
-    return _selectionChangeController.stream;
+    return _selectionChangeController!.stream;
   }
 }
 
@@ -80,9 +78,16 @@ class _SelectionChangeRecordImpl<T> extends ChangeRecord
   @override
   final Iterable<T> removed;
 
-  factory _SelectionChangeRecordImpl({Iterable<T> added, Iterable<T> removed}) {
-    added = added != null ? UnmodifiableListView(added) : const [];
-    removed = removed != null ? UnmodifiableListView(removed) : const [];
+  factory _SelectionChangeRecordImpl(
+      {Iterable<T> added = const [], Iterable<T> removed = const []}) {
+    added = UnmodifiableListView(added);
+    removed = UnmodifiableListView(removed);
+    /*    
+    added = (added != null ? UnmodifiableListView(added) : const [])
+        as Iterable<T>;
+    removed = (removed != null ? UnmodifiableListView(removed) : const [])
+        as Iterable<T>;
+    */
     return _SelectionChangeRecordImpl._(added, removed);
   }