+++ angular_components/lib/src/utils/async/priority_stream_iterator.dart
@@ -10,17 +10,17 @@ import 'package:collection/collection.dart';
 /// queue and selects the one with the most priority (the least one by
 /// comparison) of all accumulated values at each [moveNext] call.
 class PriorityStreamIterator<T extends Comparable<Object>>
-    implements StreamIterator<T> {
+    implements StreamIterator<T?> {
   final StreamIterator<T> _iterator;
   final PriorityQueue<T> _queue;
 
-  T _current;
-  Future<bool> _next;
+  T? _current;
+  late Future<bool> _next;
 
   /// Create a [PriorityStreamIterator] on [stream] with an optional
   /// [comparison] function. If [comparison] is not provided, [T] must implement
   /// Comparable<T>.
-  PriorityStreamIterator(Stream<T> stream, [int comparison(T a, T b)])
+  PriorityStreamIterator(Stream<T> stream, [int comparison(T a, T b)?])
       : _iterator = StreamIterator(stream),
         _queue = _StablePriorityQueue<T>(comparison) {
     _accumulateValues();
@@ -44,10 +44,10 @@ class PriorityStreamIterator<T extends Comparable<Object>>
   }
 
   @override
-  T get current => _current;
+  T? get current => _current;
 
   @override
-  Future<Object> cancel() {
+  Future<Object?> cancel() {
     _clear();
     return _iterator.cancel();
   }
@@ -84,10 +84,12 @@ class PriorityStreamIterator<T extends Comparable<Object>>
 /// default [Comparable.compare].
 class _StablePriorityQueue<T extends Comparable<Object>>
     extends HeapPriorityQueue<T> {
-  _StablePriorityQueue([Comparator<T> comparison])
+  _StablePriorityQueue([Comparator<T>? comparison])
       : this._(_OrderedComparator(comparison ?? _defaultComparator<T>()));
 
-  _StablePriorityQueue._(this.comparator) : super(comparator);
+  //_StablePriorityQueue._(this.comparator) : super(comparator as int Function(T*, T*)?);
+  _StablePriorityQueue._(this.comparator)
+      : super(comparator as int Function(T, T)?);
 
   final _OrderedComparator<T> comparator;
 
@@ -136,7 +138,8 @@ Comparator<T> _defaultComparator<T extends Comparable<Object>>() =>
 
 /// A [Comparator] that allows registering elements and uses the order of
 /// registration to resolve the cases when elements compare as equal.
-class _OrderedComparator<T extends Comparable<Object>> implements Function {
+//class _OrderedComparator<T extends Comparable<Object>> implements Function {
+class _OrderedComparator<T extends Comparable<Object>> {
   static const RENUMERATE_THRESHOLD = 1000000;
 
   final Comparator<T> _comparison;
@@ -177,7 +180,7 @@ class _OrderedComparator<T extends Comparable<Object>> implements Function {
       throw StateError(
           "Comparing elements that weren't registered with the comparator.");
     }
-    return _ordinalByElement[a] - _ordinalByElement[b];
+    return _ordinalByElement[a]! - _ordinalByElement[b]!;
   }
 
   void _renumerate() {